refactor: initial generalization of sidebar
