+++ src/features/nav-header/top-bar/top-bar.module.scss
@@ -13,7 +13,7 @@
   height: $topbar-height;
   font-size: 14px;
   color: $light-80;
-  z-index: 20;
+  z-index: 40;
 }
 
 .middle {
