+++ src/features/side-panel/atoms.ts
@@ -1,0 +1,7 @@
+'use client'
+
+import { atom } from 'jotai'
+
+export const headerHeightAtom = atom(0)
+
+export const hoveringCloseAtom = atom(false)
